fix(sonar): collapse duplicated Builder/legacy usage mounts
SonarCloud gated the PR on 4.2% new-code duplication from identical
Builder vs legacy usage handlers and shared OpenAPI usage query fields.
