+++ src/app/api/v1/apps/[id]/usage/route.ts
@@ -1,23 +1,2 @@
-import { NextResponse } from "next/server";
-
-import {
-  handleAppUsageGet,
-  resolveAppForUsageAccess,
-} from "@/lib/usage/app-usage-handlers";
-
 /** Legacy alias of Builder usage — M2M Basic only (provider session use Internal). */
-export async function GET(
-  request: Request,
-  { params }: { params: Promise<{ id: string }> },
-) {
-  const { id: clientId } = await params;
-  const app = await resolveAppForUsageAccess({
-    request,
-    clientId,
-    m2mOnly: true,
-  });
-  if (!app) {
-    return NextResponse.json({ error: "Not found" }, { status: 404 });
-  }
-  return handleAppUsageGet({ request, app });
-}
+export { getBuilderAppUsage as GET } from "@/lib/usage/app-usage-handlers";
