+++ src/app/api/v1/internal/dashboard/usage/route.ts
@@ -1,8 +1,2 @@
-import { NextRequest } from "next/server";
-
-import { handleDashboardUsageGet } from "@/lib/usage/dashboard-usage-handler";
-
 /** Internal alias of GET /api/v1/dashboard/usage */
-export async function GET(request: NextRequest) {
-  return handleDashboardUsageGet(request);
-}
+export { GET } from "@/app/api/v1/dashboard/usage/route";
