+++ src/app/api/v1/internal/me/usage/requests/route.ts
@@ -1,8 +1,2 @@
-import { NextRequest } from "next/server";
-
-import { handleMeUsageRequestsGet } from "@/lib/usage/me-usage-requests-handler";
-
 /** Internal alias of GET /api/v1/me/usage/requests */
-export async function GET(request: NextRequest) {
-  return handleMeUsageRequestsGet(request);
-}
+export { GET } from "@/app/api/v1/me/usage/requests/route";
